requirements.txt
fastapi==0.100.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
SQLAlchemy==2.0.19
SQLAlchemy-Utils==0.41.1
uvicorn==0.23.1
Werkzeug==2.3.7
fastapi-jwt-auth==0.5.0
pydantic==1.10.11
